+++ lib/subspace/commands/bootstrap.rb
@@ -1,6 +1,9 @@
 class Subspace::Commands::Bootstrap < Subspace::Commands::Base
+  PASS_THROUGH_PARAMS = ["private-key"]
+
   def initialize(args, options)
     @host_spec = args.first
+    @options = options
     @ask_pass = options.password
     @yum = options.yum
     run
@@ -23,6 +26,7 @@ class Subspace::Commands::Bootstrap < Subspace::Commands::Base
       "path=/home/{{ansible_ssh_user}}/.ssh state=directory mode=0700",
       "-vvvv"
     ]
+    cmd = add_pass_through_params cmd
     bootstrap_command cmd
   end
 
@@ -37,6 +41,7 @@ class Subspace::Commands::Bootstrap < Subspace::Commands::Base
       "--become",
       "-vvvv"
     ]
+    cmd = add_pass_through_params cmd
     bootstrap_command cmd
   end
 
@@ -47,4 +52,15 @@ class Subspace::Commands::Bootstrap < Subspace::Commands::Base
     ansible_command *cmd
   end
 
+  def add_pass_through_params(cmd)
+    PASS_THROUGH_PARAMS.each do |param_name|
+      x = param_name.split('-')[1..-1].map(&:upcase).join('_')
+      hash_key = (param_name.gsub('-', '_') + (x == '' ? '' : "_#{x}")).to_sym
+      value = @options.__hash__[hash_key]
+      if value
+        cmd += ["--#{param_name}", value]
+      end
+    end
+    cmd
+  end
 end